+++ ft-sdk/src/main/java/com/ft/sdk/FTTraceConfigManager.java
@@ -39,6 +39,18 @@ public class FTTraceConfigManager {
     }
 
     /**
+     * 返回自定义覆盖全局的 {@link FTTraceInterceptor}
+     *
+     * @return
+     */
+    FTTraceInterceptor.HeaderHandler getOverrideHeaderHandler() {
+        if (config == null) return null;
+        if (config.getOkHttpTraceHeaderHandler() == null) return null;
+        return config.getOkHttpTraceHeaderHandler();
+    }
+
+
+    /**
      * @return
      */
     @Nullable